+++ examples/swr/src/gen/hooks/useUpdatePetWithForm.ts
@@ -38,19 +38,19 @@ export async function updatePetWithForm(
   return res.data
 }
 
+export type UpdatePetWithFormMutationArg = { petId: UpdatePetWithFormPathParams['petId']; params?: UpdatePetWithFormQueryParams }
+
 /**
  * @summary Updates a pet in the store with form data
  * {@link /pet/:petId}
  */
 export function useUpdatePetWithForm(
-  petId: UpdatePetWithFormPathParams['petId'],
-  params?: UpdatePetWithFormQueryParams,
   options: {
     mutation?: SWRMutationConfiguration<
       UpdatePetWithFormMutationResponse,
       ResponseErrorConfig<UpdatePetWithForm405>,
       UpdatePetWithFormMutationKey | null,
-      never
+      UpdatePetWithFormMutationArg
     > & { throwOnError?: boolean }
     client?: Partial<RequestConfig> & { client?: typeof fetch }
     shouldFetch?: boolean
@@ -59,9 +59,14 @@ export function useUpdatePetWithForm(
   const { mutation: mutationOptions, client: config = {}, shouldFetch = true } = options ?? {}
   const mutationKey = updatePetWithFormMutationKey()
 
-  return useSWRMutation<UpdatePetWithFormMutationResponse, ResponseErrorConfig<UpdatePetWithForm405>, UpdatePetWithFormMutationKey | null>(
+  return useSWRMutation<
+    UpdatePetWithFormMutationResponse,
+    ResponseErrorConfig<UpdatePetWithForm405>,
+    UpdatePetWithFormMutationKey | null,
+    UpdatePetWithFormMutationArg
+  >(
     shouldFetch ? mutationKey : null,
-    async (_url) => {
+    async (_url, { arg: { petId, params } }) => {
       return updatePetWithForm(petId, params, config)
     },
     mutationOptions,