+++ packages/plugin-react-query/src/plugin.ts
@@ -36,8 +36,8 @@ export const pluginReactQuery = definePlugin<PluginReactQuery>((options) => {
     client,
   } = options
 
-  const clientType = client?.client ?? 'axios'
-  const clientImportPath = client?.importPath ?? (!client?.bundle ? `@kubb/plugin-client/clients/${clientType}` : undefined)
+  const clientName = client?.client ?? 'axios'
+  const clientImportPath = client?.importPath ?? (!client?.bundle ? `@kubb/plugin-client/clients/${clientName}` : undefined)
 
   return {
     name: pluginReactQueryName,
@@ -45,7 +45,8 @@ export const pluginReactQuery = definePlugin<PluginReactQuery>((options) => {
       output,
       client: {
         ...options.client,
-        client: clientType,
+        client: clientName,
+        clientType: client?.clientType ?? 'function',
         dataReturnType: client?.dataReturnType ?? 'data',
         pathParamsType,
         importPath: clientImportPath,
@@ -71,11 +72,14 @@ export const pluginReactQuery = definePlugin<PluginReactQuery>((options) => {
               ...query,
             },
       mutationKey,
-      mutation: {
-        methods: ['post', 'put', 'patch', 'delete'],
-        importPath: '@tanstack/react-query',
-        ...mutation,
-      },
+      mutation:
+        mutation === false
+          ? false
+          : {
+              methods: ['post', 'put', 'patch', 'delete'],
+              importPath: '@tanstack/react-query',
+              ...mutation,
+            },
       paramsType,
       pathParamsType,
       parser,