+++ packages/kubb/CHANGELOG.md
@@ -1,5 +1,53 @@
 # kubb
 
+## 3.0.154
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @kubb/cli@4.10.0
+  - @kubb/core@4.10.0
+
+## 3.0.153
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @kubb/cli@4.9.4
+  - @kubb/core@4.9.4
+
+## 3.0.152
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @kubb/cli@4.9.3
+  - @kubb/core@4.9.3
+
+## 3.0.151
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @kubb/cli@4.9.2
+  - @kubb/core@4.9.2
+
+## 3.0.150
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @kubb/cli@4.9.1
+  - @kubb/core@4.9.1
+
+## 3.0.149
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @kubb/cli@4.9.0
+  - @kubb/core@4.9.0
+
 ## 3.0.148
 
 ### Patch Changes